requirements.txt
attrs==23.1.0
certifi==2023.7.22
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
outcome==1.3.0.post0
PySocks==1.7.1
robotframework==6.1.1
robotframework-pythonlibcore==4.2.0
robotframework-seleniumlibrary==6.1.3
selenium==4.15.2
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.23.1
trio-websocket==0.11.1
urllib3==2.0.7
wsproto==1.2.0
